trying to fix mobile b ug thing calendar

+++ routes/calendarRoutes.js
@@ -5,6 +5,22 @@ const path = require("path");
 const dbWriteEvent = require("../serverscripts/calendar/writeCalendarEvent.js");
 const dbReadEvent = require("../serverscripts/calendar/readCalendarEvent.js");
 const BASE_URL = require("../BASE_URL.js");
+let lastEventID;
+
+const updateEventID = async function(){
+    await dbReadEvent.get("SELECT LAST_VALUE(id) OVER (ORDER BY id ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS lastEvent FROM events;", (err, id) => {
+        if(!id){
+            console.log("no events in database");
+        }
+        else if (err) {
+            console.log(err.message);
+            throw err;
+        } else {
+            console.log("last calendar event: " + id.lastEvent);
+            lastEventID = id.lastEvent;
+        }
+    });
+}
 
 router.get("/", (req, res) => {//homepage
     res.render("calendar", {BASE_URL: BASE_URL, username: req.session.user, permissions: req.session.permissions});
@@ -28,22 +44,9 @@ router.get("/events", async (req, res) => {//returns json of events
 
 router.get("/event/:id", async (req, res) => {
     //copypasted from legacyCalendar bc its GOOD
-    const id = req.params.id;
-    let lastEventID;
-
     try {
-        lastEventID = await new Promise((resolve, reject) => {
-            dbReadEvent.get("SELECT LAST_VALUE(id) OVER (ORDER BY id ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING) AS lastEvent FROM events;", (err, id) => {
-                if (err) {
-                    reject(err);
-                } else {
-                    resolve(id.lastEvent);
-                }
-            });
-        });
-
         const row = await new Promise((resolve, reject) => {
-            dbReadEvent.get("SELECT * FROM events WHERE id = ?", [id], (err, row) => {
+            dbReadEvent.get("SELECT * FROM events WHERE id = ?", [req.params.id], (err, row) => {
                 if (err) {
                     reject(err);
                 } else {
@@ -51,7 +54,8 @@ router.get("/event/:id", async (req, res) => {
                 }
             });
         });
-        res.render("./calendar/calendarday" ,{ events: row , lastEventID: lastEventID, BASE_URL: process.env.BASE_URL});
+        if(!lastEventID) await updateEventID();
+        res.render("./calendar/calendarday" ,{ events: row , lastEventID: lastEventID, BASE_URL: BASE_URL});
     } catch (err) {
         res.status(500).json({ error: err.message });
     }
@@ -76,6 +80,8 @@ router.post("/submitEvent", async (req, res) => {
             }
             else{
                 console.log(title + " event creation success");
+                lastEventID++;
+                console.log("new last calendar event id: " + lastEventID);
                 res.status(200).json({ message: title + " event creation success" });
             }
         });
@@ -88,16 +94,54 @@ router.post("/submitEvent", async (req, res) => {
 router.post("/deleteEvent", async (req, res) => {
     if(req.session.permissions === "admin"){
         const id = req.body.id;
-        dbWriteEvent.run("DELETE FROM events WHERE id = ?", [id], (err) => {
+        dbWriteEvent.get("SELECT * FROM events WHERE id = (?)", [id], (err, row) => {
             if(err){
-                console.log(err);
-                res.status(500).json({ error: err.message });
+                console.error(err.message);
+                res.status(500).json({ message: "ERROR", error: err.message });
+            }
+            else if(row){
+
+                dbWriteEvent.serialize(() => {
+                    console.log("event deletion attempt")
+                    dbWriteEvent.run("BEGIN TRANSACTION");
+                    dbWriteEvent.run("DELETE FROM events WHERE id = ?", [id], (err) => {
+                        if(err){
+                            console.log(err);
+                            res.status(500).json({ error: err.message });
+                        }
+                        else{
+                            console.log("deletion success");
+                            lastEventID--;
+                            console.log("new last calendar event id: " + lastEventID);
+                            res.status(200).json({message: `id ${id} deletion success`});
+                        }
+                    });
+                    
+                    
+                    dbWriteEvent.run(`CREATE TABLE events_temp(
+                        id INTEGER PRIMARY KEY AUTOINCREMENT,
+                        user TEXT NOT NULL,
+                        title TEXT NOT NULL,
+                        location TEXT NOT NULL,
+                        description TEXT,
+                        date TEXT NOT NULL,
+                        starttime TEXT NOT NULL,
+                        endtime TEXT
+                    )`);
+
+                    dbWriteEvent.run("INSERT INTO events_temp(user, title, location, description, date, starttime, endtime) SELECT user, title, location, description, date, starttime, endtime FROM events;");
+                    dbWriteEvent.run("DROP TABLE events;");
+                    dbWriteEvent.run("ALTER TABLE events_temp RENAME TO events;");    
+    
+                    dbWriteEvent.run("COMMIT");
+                });
             }
             else{
-                console.log("deletion success");
-                res.status(200).json({message: "deletion success"});
+                console.log("event does not exist");
+                res.status(404).json({message: "ERROR", error: "wher event"});
             }
         });
+        
     }
     else{
         console.log("non-admin deletion attempt");
